Changed this files during and immediatly after meeting with teacher

+++ exercicios/aula1/ex1f.py
@@ -1,4 +1,4 @@
-lado = int(input("Insira o lado do quadrado: "))
+lado = float(input("Insira o lado do quadrado: "))
 
 print(f"A área do quadrado é {lado * lado}")
 print(f"O perimetro do quadrado é {lado * 4}")
